Extracted the generic CLI-argument framework into PortableUtil
* Moved Argument/ArgumentBase/NameValuePair/ArgumentGroup/ParaUsageBlock/IUsageBlock
  and the base + 10 context-free value exceptions to pwiz.Common.CommandLine, generified
  on TContext (Argument<CommandArgs> in Skyline via file-scoped using aliases)
* Decoupled host text/behavior via ArgUsage seams (IArgUsageProvider descriptions/headers/
  error messages, IsRemoteUrl, IsWrappableListType, HtmlEncode); Skyline installs them in a
  static ctor, keeping usage HTML byte-identical (HttpUtility encoder retained)
* Kept Skyline-domain value exceptions subclassing the moved bases; updated test/CommandLine
  references and added PortableUtil project references

See ai/todos/active/TODO-20260622_portableutil_cli_args_framework.md
