+++ pwiz_tools/Shared/PortableUtil/CommandLine/IUsageBlock.cs
@@ -1,0 +1,31 @@
+/*
+ * Original author: Brendan MacLean <brendanx .at. u.washington.edu>,
+ *                  MacCoss Lab, Department of Genome Sciences, UW
+ *
+ * Copyright 2026 University of Washington - Seattle, WA
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace pwiz.Common.CommandLine
+{
+    /// <summary>
+    /// A renderable section of command-line usage output (an argument group or a
+    /// paragraph). Rendered as plain text (ascii/unicode tables) or as HTML.
+    /// </summary>
+    public interface IUsageBlock
+    {
+        string ToString(int width, string formatType);
+        string ToHtmlString();
+    }
+}
