+++ pwiz_tools/Skyline/TestData/CommandLineRefineTest.cs
@@ -24,6 +24,8 @@ using System.IO;
 using System.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using pwiz.Skyline;
+using pwiz.Common.CommandLine;
+using Argument = pwiz.Common.CommandLine.Argument<pwiz.Skyline.CommandArgs>;
 using pwiz.Skyline.Model;
 using pwiz.Skyline.Model.AuditLog;
 using pwiz.Skyline.Model.GroupComparison;
@@ -433,7 +435,7 @@ namespace pwiz.SkylineTestData
         [TestMethod]
         public void ConsoleArgumentInvalidValuesTest()
         {
-            var testedArguments = new HashSet<CommandArgs.Argument>
+            var testedArguments = new HashSet<Argument>
             {
                 CommandArgs.ARG_PANORAMA_FOLDER,
                 CommandArgs.ARG_TOOL_INITIAL_DIR,
@@ -443,7 +445,7 @@ namespace pwiz.SkylineTestData
                 CommandArgs.ARG_PEPTIDE_ADD_MOD_AA, // tested in ConsoleModsTest
                 CommandArgs.ARG_PEPTIDE_ADD_MOD_TERM, // tested in ConsoleModsTest
             };
-            var allArgumentsSet = new HashSet<CommandArgs.Argument>(
+            var allArgumentsSet = new HashSet<Argument>(
                 CommandArgs.AllArguments.Where(a => !a.InternalUse && !testedArguments.Contains(a)));
 
             TestFilesDir = new TestFilesDir(TestContext, @"TestData\CommandLineRefine.zip");
@@ -468,18 +470,18 @@ namespace pwiz.SkylineTestData
             Assert.AreEqual(0, allArgumentsSet.Count, string.Join(", ", allArgumentsSet.Select(a => a.Name)));
         }
 
-        private void ValidateInvalidValue(CommandArgs.Argument arg, HashSet<CommandArgs.Argument> testedArguments)
+        private void ValidateInvalidValue(Argument arg, HashSet<Argument> testedArguments)
         {
             const string NO_VALUE = "NO VALUE";
             string expected = string.Format(
                 CommandArgUsage.ValueInvalidException_ValueInvalidException_The_value___0___is_not_valid_for_the_argument__1___Use_one_of__2_,
                 NO_VALUE, arg.ArgumentText, arg.ValueExample());
             expected = expected.Substring(0, expected.IndexOf(@"None, ", StringComparison.Ordinal) + 6);
-            AssertEx.ThrowsException<CommandArgs.ValueInvalidException>(() => arg.GetArgumentTextWithValue(NO_VALUE), expected);
+            AssertEx.ThrowsException<ValueInvalidException>(() => arg.GetArgumentTextWithValue(NO_VALUE), expected);
             testedArguments.Add(arg);
         }
 
-        private void ValidateInvalidValueNumeric(CommandArgs.Argument arg, HashSet<CommandArgs.Argument> testedArguments)
+        private void ValidateInvalidValueNumeric(Argument arg, HashSet<Argument> testedArguments)
         {
 
             const string BAD_VALUE = "-1a";
@@ -509,10 +511,10 @@ namespace pwiz.SkylineTestData
             testedArguments.Add(arg);
         }
 
-        private void ValidateInvalidValueBool(CommandArgs.Argument arg, HashSet<CommandArgs.Argument> testedArguments)
+        private void ValidateInvalidValueBool(Argument arg, HashSet<Argument> testedArguments)
         {
             const string BAD_VALUE = "-1a";
-            AssertEx.ThrowsException<CommandArgs.ValueUnexpectedException>(() => arg.GetArgumentTextWithValue(BAD_VALUE), arg.ArgumentText);
+            AssertEx.ThrowsException<ValueUnexpectedException>(() => arg.GetArgumentTextWithValue(BAD_VALUE), arg.ArgumentText);
             var args = new[] { "--in=" + DocumentPath, "--overwrite", arg.ArgumentText + "=BAD_VALUE" };
             string output = RunCommand(args);
             string expected = string.Format(Resources.ValueUnexpectedException_ValueUnexpectedException_The_argument__0__should_not_have_a_value_specified, arg.ArgumentText);
@@ -520,7 +522,7 @@ namespace pwiz.SkylineTestData
             testedArguments.Add(arg);
         }
 
-        private void ValidateInvalidValueString(CommandArgs.Argument arg, HashSet<CommandArgs.Argument> testedArguments)
+        private void ValidateInvalidValueString(Argument arg, HashSet<Argument> testedArguments)
         {
             var args = new[] { "--in=" + DocumentPath, "--overwrite", arg.ArgumentText /* no value */ };
             string output = RunCommand(args);
@@ -529,7 +531,7 @@ namespace pwiz.SkylineTestData
             testedArguments.Add(arg);
         }
 
-        private void ValidateInvalidValuePath(CommandArgs.Argument arg, HashSet<CommandArgs.Argument> testedArguments)
+        private void ValidateInvalidValuePath(Argument arg, HashSet<Argument> testedArguments)
         {
             const string BAD_VALUE = "Bad:\\Path\\Value";
             string argText = arg.GetArgumentTextWithValue(BAD_VALUE);
@@ -583,11 +585,11 @@ namespace pwiz.SkylineTestData
             AssertEx.Contains(output, new CommandArgs.ValueInvalidIonTypeListException(CommandArgs.ARG_TRAN_FRAGMENT_ION_TYPES, typesWithError).Message);
             const int outOfRangeCharge = Transition.MAX_PRODUCT_CHARGE + 1;
             output = Run(CommandArgs.ARG_TRAN_FRAGMENT_ION_CHARGES.GetArgumentTextWithValue(outOfRangeCharge.ToString()));
-            AssertEx.Contains(output, new CommandArgs.ValueOutOfRangeIntException(CommandArgs.ARG_TRAN_FRAGMENT_ION_CHARGES, outOfRangeCharge,
+            AssertEx.Contains(output, new ValueOutOfRangeIntException(CommandArgs.ARG_TRAN_FRAGMENT_ION_CHARGES, outOfRangeCharge,
                 Transition.MIN_PRODUCT_CHARGE, Transition.MAX_PRODUCT_CHARGE).Message);
             const int outOfRangePrecursorCharge = TransitionGroup.MAX_PRECURSOR_CHARGE + 1;
             output = Run(CommandArgs.ARG_TRAN_PRECURSOR_ION_CHARGES.GetArgumentTextWithValue("1, 2, " + outOfRangePrecursorCharge));
-            AssertEx.Contains(output, new CommandArgs.ValueOutOfRangeIntException(CommandArgs.ARG_TRAN_PRECURSOR_ION_CHARGES, outOfRangePrecursorCharge,
+            AssertEx.Contains(output, new ValueOutOfRangeIntException(CommandArgs.ARG_TRAN_PRECURSOR_ION_CHARGES, outOfRangePrecursorCharge,
                 TransitionGroup.MIN_PRECURSOR_CHARGE, TransitionGroup.MAX_PRECURSOR_CHARGE).Message);
             const string chargesWithError = "2, 3, p, 5";
             output = Run(CommandArgs.ARG_TRAN_PRECURSOR_ION_CHARGES.GetArgumentTextWithValue(chargesWithError));
